Refactor parking trash pages directories
